+++ docs/main/changelog.rst
@@ -7,6 +7,10 @@ that were made in every particular version.
 From version 0.7.6 *Dependency Injector* framework strictly 
 follows `Semantic versioning`_
 
+Development version
+-------------------
+- Add new example miniapp "ioc_container".
+
 3.12.0
 ------
 - Regenerate C sources using Cython 0.28.2.